stability

+++ source/engine/texture.h
@@ -61,8 +61,8 @@ struct TextureOptions {
 /* Represents an actual texture on the GPU */
 struct Texture {
     sg_image id; /* ID reference for the GPU memory location of the texture */
-    int width;
-    int height;
+    uint16_t width;
+    uint16_t height;
     unsigned char *data;
     struct TextureOptions opts;
     struct TexAnim anim;
